refactor: split timelog parser by action family
